requirements.txt
Django>=1.11,<1.12
django-bootstrap3>=11.0.0
hashids>=1.2.0
MySQL-python>=1.2.5
pyquery>=1.2.13,!=1.2.14
factory_boy>=2.11
